requirements.txt
# requirements for Django backend
# Incluye paquetes comunes para REST API, autenticación, CORS, Postgres, tareas asíncronas y manejo de archivos estáticos.

Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=3.14
psycopg2-binary>=2.9
python-dotenv>=1.0
django-environ>=0.10
Pillow>=9.0
django-filter>=23.0
drf-spectacular>=0.26
whitenoise>=6.4
gunicorn>=20.1.0
celery>=5.2
redis>=4.5

# Opcionales (descomentar si los necesita)
# channels>=4.0    # websockets / Django Channels
# django-debug-toolbar>=4.0
# coverage>=6.0
